+++ src/main/java/rtg/proxy/ClientProxy.java
@@ -1,8 +1,13 @@
 package rtg.proxy;
 
+import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.fml.common.event.FMLInitializationEvent;
 import net.minecraftforge.fml.common.event.FMLPostInitializationEvent;
 import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
+
+import rtg.event.GuiEventHandler;
+import rtg.event.WorldTypeMessageEventHandler;
+import static rtg.api.RTGAPI.config;
 
 public class ClientProxy extends CommonProxy
 {
@@ -10,7 +15,6 @@ public class ClientProxy extends CommonProxy
 
     @Override
     public void preInit(FMLPreInitializationEvent event) {
-
         super.preInit(event);
     }
 
@@ -18,11 +22,18 @@ public class ClientProxy extends CommonProxy
     public void init(FMLInitializationEvent event) {
 
         super.init(event);
+
+        if (config().ENABLE_WORLD_TYPE_NOTIFICATION_SCREEN.get()) {
+            MinecraftForge.EVENT_BUS.register(WorldTypeMessageEventHandler.instance);
+        }
+
+        if (config().USE_RTG_WORLD_TYPE_DEFAULT.get()) {
+            MinecraftForge.EVENT_BUS.register(GuiEventHandler.instance);
+        }
     }
 
     @Override
     public void postInit(FMLPostInitializationEvent event) {
-
         super.postInit(event);
     }
 }
